hrr, int promotion

+++ trunk/src/blob.c
@@ -319,7 +319,7 @@ fmt_H(int pack, int len, blob_t *b, va_list *ap)
 	if (len) return (-1);
 	
 	if (pack) {
-		uint16_t n = va_arg(*ap, uint16_t);
+		uint16_t n = va_arg(*ap, int);
 		n = htons(n);
 		if (blob_write(b, &n, sizeof(n)) < 0)
 			return (-1);
@@ -351,7 +351,7 @@ fmt_c(int pack, int len, blob_t *b, va_list *ap)
 	if (len) return (-1);
 	
 	if (pack) {
-		uint8_t n = va_arg(*ap, uint8_t);
+		uint8_t n = va_arg(*ap, int);
 		return (blob_write(b, &n, sizeof(n)));
 	} else {
 		uint8_t *n = va_arg(*ap, uint8_t *);
@@ -379,7 +379,7 @@ fmt_h(int pack, int len, blob_t *b, va_list *ap)
 	if (len) return (-1);
 	
 	if (pack) {
-		uint16_t n = va_arg(*ap, uint16_t);
+		uint16_t n = va_arg(*ap, int);
 		return (blob_write(b, &n, sizeof(n)));
 	} else {
 		uint16_t *n = va_arg(*ap, uint16_t *);
